Preliminary recognition of Windows 10 Version 1709 (Fall Creators update) and future build ranges.
Just prior to release of a feature update, a stablization branch is created for the update. As of September 2017, this branch is rs3_release for Version 1709 (Fall Creators Update). For convenience, recognize the first slow ring build from this branch (build 16278) as the minimum build candidate range for version 1709. For future releases, catch build candidate ranges and announce it as the next release ID.