Windows Insier detection: use self-host key to detect if Insider Preview is in use.
In September 2019, 20H1 builds have moved to 'vb_release' branch, causing NVDA to say 'Ver1903' instead of 'Insider'. Turns out there is a more reliable way to check whether Insider Preview build is in use: Windows self-host key/Applicability/IsRetailOS. If this is yes or not present, it is a public build, otherwise it is an Insider build. Note that this key must be viewed from native Registry view, so 64-bit access must be used for NVDA because it is a 32-bit app.